Bootstrapped State Fractions

The function subsamples from the rows (observations / cells) in <scores> and uses the subsetted matrix to compute state fractions. This process is repeated <n.iter> times with subsamples of <sample.size>.

bootstrapped_state_fractions(
  scores,
  sample.size = 200,
  n.iter = 10000L,
  min = NULL,
  diff = NULL,
  summary = F
)

Arguments

scores

a dataframe of cell signature scores. Cells are rows and signatures are columns.

sample.size

the size of the subsample to be taken in each bootstrap. Default: 200

n.iter

the number of bootstraps or iterations. Default: 10000

min

minimum value required for a row to be assigned to a column. Default: NULL

diff

minimum difference in value to the 'next-best' row that is required for a row to be assigned to a column. Default: NULL

summary

logical; return summary statistics for bootstraps instead of the raw data. See `?scalop::bootstrap_summary` for details. Default: F

Value

a dataframe with as many columns as <n.iter> and as many rows as the number of columns (signatures) in <scores>. if summary = TRUE, instead returns a dataframe with two columns: mean and SEM.