Convert a list to a jaccard matrix and plot with `scalop::gmap`

Convert a list to a jaccard matrix and plot with `scalop::gmap`

jacmap(
  L = NULL,
  L2 = NULL,
  m = NULL,
  ratio = 1,
  limits = c(0, 1),
  tile.col = "black",
  tile.size = 0.1,
  cols = c("white", "darkred"),
  num = FALSE,
  angle = T,
  ...,
  tiletext = FALSE,
  dcp = 1
)

Arguments

L

list of character vectors to compute jaccard indexes between.

ratio

aspect ratio. Default: 1

limits

colour key limits. Default: c(0, 1)

tile.col

draw lines between tiles. Default: 'black'

tile.size

draw lines between tiles. Default: 0.1

cols

colours for key. Default: c("white", "darkred")

num

numeric (TRUE) or character (FALSE) x- and y- axes labels. Default: FALSE

angle

angle x-axis labels. Default: T

...

other arguments passed to `scalop::gmap`

tiletext

write jaccard indexes on tiles? Default: FALSE

dcp

relevant if tiletext is TRUE. How many decimal places to show? Default: 1

Value

ggplot object

See also