Compute Metaprograms

programs and profiles are the lists of gene sets and DE profiles that were derived from intra-tumour cell clusters. These are integrated, according to the metaclusters clustering, into a final list of metaprograms; one per metacluster.

metaprograms(
  programs,
  profiles,
  m,
  metaclusters,
  samples = NULL,
  freq.cutoff = 3,
  order.priority = c("freq", "lfc"),
  return.data = FALSE
)

Arguments

programs

list of differentially expressed genes in each group/cell cluster

profiles

list of differential expression profiles for all genes in each group/cell cluster

m

matrix of genes by cells

metaclusters

list of program metaclusters

samples

character vector of unique sample names. Default: readRDS("samples_2020-03-01.rds")

order.priority

order by 'freq' first or 'lfc' first. Default: 'freq'

freq

minimum number of metacluster-specific sample-specific programs a gene must appear in. Default: 3

Value

if return.data = F, just the list of metaprograms. Else, the list of metaprograms, and the dataframe of metacluster-, sample- and gene- specific LFC and freq values.

Details

DETAILS

See also