Principal Component Analysis

A wrapper for irlba::prcomp_irlba(). Performs efficient computation of a principal component analysis of a given data matrix.

pca(m, npcs = 100, retx = TRUE, center = TRUE, scale = FALSE)

Arguments

m

A numeric matrix or data frame.

npcs

The number of principal component vectors to return. Must be less than 'min(dim(m))' Default: 100

retx

A logical value indicating whether the rotated variables should be returned. Default: TRUE

center

A logical value indicating whether the variables should be shifted to be zero centered. Alternately, a centering vector of length equal to the number of columns of 'm' can be supplied. Default: TRUE

scale

A logical value indicating whether the variables should be scaled to have unit variance before the analysis is performed. The value of 'scale' determines how column scaling is performed (after centering). If 'scale' is a numeric vector with length equal to the number of columns of 'm', then each column of 'm' is divided by the corresponding value from 'scale'. If 'scale' is 'TRUE', then scaling is done by dividing the (centered) columns of 'm' by their standard deviations if 'center=TRUE', and the root mean square otherwise. If 'scale' is 'FALSE', no scaling is done. Default: FALSE

Value

A list of class "prcomp" containing the following components:

sdev: the standard deviations of the principal components (i.e. the square roots of the eigenvalues of the covariance/correlation matrix, though the calculation is actually done with the singular values of the data matrix).

rotation: the matrix of variable loadings (i.e. a matrix whose columns contain the eigenvectors).

x: if 'retx' is 'TRUE' the value of the rotated data (the centered (and scaled if requested) data multipleied by the 'rotation' matrix) is returned. Hence, 'cov(m)' is the diagonal matrix 'diag(sdev^2)'.

center, scale: the centering and scaling used, or 'FALSE'.

Details

see 'irlba::prcomp_irlba()' for more details.

See also